@@ -12,9 +12,9 @@
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.21.1",
     "cra-template": "1.0.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-scripts": "^3.4.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "predeploy": "npm run build",
