@@ -1,17 +1,17 @@
-import React, { useEffect, useMemo, useState } from 'react';
-import { API_KEY } from '../../secrets.json';
+import React, { useEffect, useId, useState } from 'react';
 import { useLocalStorage } from '../../hooks/useLocalStorage';
 import { TopBar } from '../top-bar/TopBar';
 import { MainSection } from '../main-section/MainSection';
 import { SettingsProvider } from '../../context/settings-context';
 
 export function App() {
-  const trendingGifsQueryCode = useMemo(() => String(Math.random()), []);
+  const trendingGifsQueryCode = useId();
   const [gifs, setGifs] = useState([]);
   const [queryString, setQueryString] = useState(trendingGifsQueryCode);
   const [apiResOffset, setApiResOffset] = useState(0);
   const [isLoading, setIsLoading] = useState(false);
   const [failedToLoad, setFailedToLoad] = useState(false);
+  const [allGifsFetched, setAllGifsFetched] = useState(false);
   const [topBarIsStyled, setTopBarIsStyled] = useState(false);
   const [infiniteScrollIsActive, setInfiniteScrollIsActive] = useLocalStorage('infiniteScroll', true);
 
@@ -22,20 +22,25 @@ export function App() {
       return;
     }
     setIsLoading(true);
+    setAllGifsFetched(false);
     const searchForTrending = query === trendingGifsQueryCode;
     const path = searchForTrending ? 'trending' : 'search';
     const q = searchForTrending ? '' : `&q=${query}`;
     // trending path finds currently trending gifs instead of searching for gifs about trending:
-    const url = `https://api.giphy.com/v1/gifs/${path}?api_key=${API_KEY}${q}&limit=${gifsPerRequest}&offset=${offset}`;
+    const url = `https://api.giphy.com/v1/gifs/${path}?api_key=${process.env.REACT_APP_GIPHY_API_KEY}${q}&limit=${gifsPerRequest}&offset=${offset}`;
     const response = await fetch(url);
-    const { data, meta } = await response.json();
+    const { data, meta: { status }, pagination: { total_count: totalCount } } = await response.json();
     setIsLoading(false);
-    const statusNotOk = meta.status < 200 || meta.status > 299;
-    if (!statusNotOk && data.length) {
-      setGifs(prev => infiniteScrollIsActive ? prev.concat(data) : data);
-      setApiResOffset(offset + gifsPerRequest);
-    } else {
+    if (status < 200 || status > 299) {
       setFailedToLoad(true);
+    } else if (data.length) {
+      setGifs(prev => infiniteScrollIsActive ? prev.concat(data) : data);
+      const numberOfNewGifs = data.length === gifsPerRequest ? gifsPerRequest : Math.min(data.length, gifsPerRequest);
+      const newOffset = offset + numberOfNewGifs;
+      setApiResOffset(newOffset);
+      setAllGifsFetched(newOffset >= totalCount);
+    } else {
+      setAllGifsFetched(true);
     }
   };
 
@@ -43,14 +48,7 @@ export function App() {
     fetchGifs();
   };
 
-  const handleLoading = () => {
-    if (isLoading && gifs.length) {
-      setIsLoading(false);
-    }
-  };
-
   useEffect(fetchGifsOnMount, []);
-  useEffect(handleLoading, [isLoading, gifs.length]);
 
   return (
     <SettingsProvider>
@@ -72,6 +70,7 @@ export function App() {
         fetchGifs={fetchGifs}
         gifsPerRequest={gifsPerRequest}
         failedToLoad={failedToLoad}
+        allGifsFetched={allGifsFetched}
         isLoading={isLoading}
         apiResOffset={apiResOffset}
         setTopBarIsStyled={setTopBarIsStyled}
